Added invalid id check to comment delete and associated unit test

+++ src/main/java/com/jaehyune/stamp_app/service/impl/CommentServiceImpl.java
@@ -91,6 +91,9 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public void delete(Integer id) {
+        if (id < 0) {
+            throw new RuntimeException("Invalid ID: " + id);
+        }
         commentRepository.deleteById(id);
     }
